SOC-1 Process Control/requirements.txt
Flask
Werkzeug
requests
python-docx
openpyxl
pandas
numpy
pytesseract
Pillow
PyMuPDF
pdf2image
pymongo
python-dotenv